+++ lib/gir_ffi-base/gobject/lib.rb
@@ -1,13 +1,6 @@
 # frozen_string_literal: true
 require 'ffi/bit_masks'
 
-# NOTE: Monkey-patch BitMask to work on JRuby.
-FFI::BitMasks::BitMask.class_eval do
-  def reference_required?
-    false
-  end
-end
-
 module GObject
   # Module for attaching functions from the gobject library
   module Lib